refactor OsgiClassnameMapper to be static and in util.core.xstream package
