+++ core/src/main/java/org/apache/brooklyn/util/core/xstream/ClassRenamingMapper.java
@@ -22,7 +22,6 @@ import static com.google.common.base.Preconditions.checkNotNull;
 
 import java.util.Map;
 
-import org.apache.brooklyn.core.mgmt.persist.OsgiClassPrefixer;
 import org.apache.brooklyn.util.guava.Maybe;
 import org.apache.brooklyn.util.javalang.Reflections;
 import org.slf4j.Logger;
@@ -39,10 +38,10 @@ import com.thoughtworks.xstream.mapper.MapperWrapper;
 public class ClassRenamingMapper extends MapperWrapper {
     
     /*
-     * TODO There is a strange relationship between this and XmlMementoSerializer$OsgiClassnameMapper.
+     * TODO There is a strange relationship between this and OsgiClassnameMapper.
      * Should these be perhaps merged?
      * 
-     * TODO For class-loading on deserialzation, should we push the class-rename logic into 
+     * TODO For class-loading on deserialization, should we push the class-rename logic into 
      * org.apache.brooklyn.util.core.ClassLoaderUtils instead? Does the xstream mapper do
      * anything else important, beyond that class-loading responsibility? It's registration
      * in XmlSerializer makes it look a bit scary: wrapMapperForAllLowLevelMentions().
