+++ core/src/main/java/org/apache/brooklyn/util/core/xstream/OsgiClassPrefixer.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.brooklyn.core.mgmt.persist;
+package org.apache.brooklyn.util.core.xstream;
 
 import javax.annotation.Nullable;
