+++ core/src/test/java/org/apache/brooklyn/util/core/xstream/OsgiClassPrefixerTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.brooklyn.core.mgmt.persist;
+package org.apache.brooklyn.util.core.xstream;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
